Two fixes for Melodic.
In p2os_driver, the p2os_msgs_gencpp doesn't exist (and, as
far as I can tell, never existed).  It worked up until recently
because older versions of CMake would complain but not fail,
while more recent cmake versions actually fail.  Just remove
the dependency.

The second fix is a cosmetic issue, where CMake would spew a
bunch of warnings on the command line.  Increase the version
to 3.9.5 to match other CMakeLists.txt and to get rid of the
warnings.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>